working solution

+++ src/webots/wren/WbWrenLabelOverlay.hpp
@@ -41,11 +41,11 @@ public:
   static void colorToArray(float *dest, int color);
 
   void setText(const QString &text);
-  void setPosition(float x, float y) {
+  void setPosition(double x, double y) {
     mX = x;
     mY = y;
   }
-  void setSize(float size) { mSize = size; }
+  void setSize(double size) { mSize = size; }
   void setColor(int color) {
     colorToArray(mColor, color);
     mTextNeedRedraw = true;
@@ -55,11 +55,11 @@ public:
   int id() const { return mId; };
   const QString &text() const { return mText; }
   const QString &font() const { return mFontName; }
-  float size() const { return mSize; }
-  float x() const { return mX; }
-  float y() const { return mY; }
+  double size() const { return mSize; }
+  double x() const { return mX; }
+  double y() const { return mY; }
   const float *color() const { return mColor; }
-  void position(float &x, float &y) const {
+  void position(double &x, double &y) const {
     x = mX;
     y = mY;
   }
@@ -99,9 +99,9 @@ private:
   int mId;
   QString mText;
   QString mFontName;
-  float mX;
-  float mY;
-  float mSize;
+  double mX;
+  double mY;
+  double mSize;
   float mColor[4];
   float mBackgroundColor[4];
   int mLinesCount;
